+++ openghg/store/_eulerian_model.py
@@ -212,18 +212,13 @@ class EulerianModel(BaseStore):
                 key: value for key, value in fn_input_parameters.items() if key not in matched_keys
             }
 
+            # Check to ensure no required keys are being passed through optional_metadata dict
+            self.check_info_keys(optional_metadata)
             if optional_metadata is not None:
                 additional_metadata.update(optional_metadata)
 
             # Mop up and add additional keys to metadata which weren't passed to the parser
             model_data = self.update_metadata(model_data, additional_input_parameters, additional_metadata)
-
-            # Use config and latest metadata to create lookup keys
-            lookup_keys = self.get_lookup_keys(optional_metadata)
-
-            if optional_metadata is not None:
-                for parsed_data in model_data.values():
-                    parsed_data["metadata"].update(optional_metadata)
 
             data_type = "eulerian_model"
             datasource_uuids = self.assign_data(
@@ -231,7 +226,6 @@ class EulerianModel(BaseStore):
                 if_exists=if_exists,
                 new_version=new_version,
                 data_type=data_type,
-                required_keys=lookup_keys,
                 compressor=compressor,
                 filters=filters,
             )