+++ openghg/store/_flux.py
@@ -206,33 +206,13 @@ class Flux(BaseStore):
             em_data = split_data["data"]
             Flux.validate_data(em_data)
 
-        # combine metadata and get look-up keys
-        if optional_metadata is None:
-            optional_metadata = {}
-
-        # TODO: REMOVE ONCE OPTIONAL KEYS CAN BE TAKEN FROM OTHER INPUTS THAN OPTIONAL_METADATA
-        optional_keywords: dict[Any, Any] = {
-            "database": database,
-            "database_version": database_version,
-            "model": model,
-        }
-
-        to_add = {key: value for key, value in optional_keywords.items() if value is not None}
-        optional_metadata.update(to_add)
-        ######
-
         # Check to ensure no required keys are being passed through optional_metadata dict
+        self.check_info_keys(optional_metadata)
         if optional_metadata is not None:
             additional_metadata.update(optional_metadata)
 
         # Mop up and add additional keys to metadata which weren't passed to the parser
         flux_data = self.update_metadata(flux_data, additional_input_parameters, additional_metadata)
-
-        lookup_keys = self.get_lookup_keys(optional_metadata=optional_metadata)
-
-        # add optional metdata to parsed metadata
-        for parsed_data in flux_data.values():
-            parsed_data["metadata"].update(optional_metadata)
 
         data_type = "flux"
         datasource_uuids = self.assign_data(
@@ -240,7 +220,6 @@ class Flux(BaseStore):
             if_exists=if_exists,
             new_version=new_version,
             data_type=data_type,
-            required_keys=lookup_keys,
             compressor=compressor,
             filters=filters,
         )