+++ openghg/store/_obscolumn.py
@@ -179,17 +179,13 @@ class ObsColumn(BaseStore):
         # this could be "site" or "satellite" keys.
         # platform = list(obs_data.keys())[0]["metadata"]["platform"]
 
+        # Check to ensure no required keys are being passed through optional_metadata dict
+        self.check_info_keys(optional_metadata)
         if optional_metadata is not None:
             additional_metadata.update(optional_metadata)
 
         # Mop up and add additional keys to metadata which weren't passed to the parser
         obs_data = self.update_metadata(obs_data, additional_input_parameters, additional_metadata)
-
-        lookup_keys = self.get_lookup_keys(optional_metadata)
-
-        if optional_metadata is not None:
-            for parsed_data in obs_data.values():
-                parsed_data["metadata"].update(optional_metadata)
 
         data_type = "column"
         datasource_uuids = self.assign_data(
@@ -197,7 +193,6 @@ class ObsColumn(BaseStore):
             if_exists=if_exists,
             new_version=new_version,
             data_type=data_type,
-            required_keys=lookup_keys,
             compressor=compressor,
             filters=filters,
         )